Completion
- Completed the file copying functionality

Program now finds the files in both directories and generates a list of the missing and new files.
Missing files are present in the left directory but not the right directory.
New files are present in the right directory but not hte left directory.
Supports copying files from the left directory to the right directory.
Supports resetting the database (by the user) to be able to do multiple checks on the same two paths.
- Auto reset occurs when a new path is selected for either the left or right directories.
Ready to be packaged! :)